RESTfulBuild/login_service/requirements.txt
uvicorn
fastapi 
python-multipart
sqlalchemy
asyncpg
PyJWT
aiohttp
pydantic
cryptography
redis